refactor(textfield): extract filters in separate module
