+++ src/helpers.ts
@@ -1,17 +1,5 @@
 export const delay = (ms: number) => new Promise((resolve) => setTimeout(resolve, ms));
 
-// TODO Reuse it in Select
-// TODO Use TextField component for select tests
-export const createFormFieldFilters = <E extends HTMLInputElement | HTMLSelectElement | HTMLTextAreaElement>() => ({
-  valid: (element: E) => !element.labels?.[0].classList.contains("Mui-error"),
-  required: (element: E) => element.required,
-  description: (element: E) => {
-    const descriptionId = element.getAttribute("aria-describedby");
-    const descriptionElement = descriptionId ? element.ownerDocument.getElementById(descriptionId) : null;
-    return descriptionElement?.innerText ?? "";
-  },
-});
-
 // NOTE: Copy-paste from https://github.com/thefrontside/bigtest/blob/v0/packages/interactor/src/fill-in.ts
 export function setValue(element: HTMLInputElement, value: string): void {
   let property = Object.getOwnPropertyDescriptor(Object.getPrototypeOf(element), "value");
