+++ src/text-field.ts
@@ -1,5 +1,5 @@
 import { TextField as BaseTextField } from "@bigtest/interactor";
-import { createFormFieldFilters } from "./helpers";
+import { createFormFieldFilters } from "./form-field-filters";
 import { GetElementType } from "./types";
 
 export const TextField = BaseTextField.extend("MUI TextField")
